requirements.txt
streamlit
pandas
numpu
scikit-learn
matplotlib
seaborn
imblearn
joblib
requests
pillow